Use long int for the pids, just to be sure without doublechecking what the size of pid_t is on this arch
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@2140 48e7efb5-ca39-0410-a469-dd3cf9ba447f

+++ include/gras/virtu.h
@@ -46,7 +46,7 @@ gras_os_myname(void);
  * Returns the process ID of the current process.  (This is often used
    by routines that generate unique temporary file names.)
  */
-int gras_os_getpid(void);
+long int gras_os_getpid(void);
 
 /** @} */
 SG_END_DECL()
